translator/pyproject.toml
[tool.poetry]
name = "translator"
version = "0.1.0"
description = ""
authors = ["yin1991 <84140478+xiaokuili@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain-openai = "^0.1.8"
langchain = "^0.2.5"
pytest = "^8.2.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
